Fix packet parsing correctness and robustness findings
Replaces dead-code exception handling with explicit Enum.IsDefined checks
for platform and session-type fallbacks, and cuts driver names at the
first null terminator instead of trimming the full fixed-size buffer.
Widens several inconsistently typed uint8 spec fields to ushort, retypes
track id to a signed short so -1 (unknown) survives, pre-populates
per-car/per-lap array elements in constructors that previously left them
null, and gives non-nullable string DTO properties explicit empty
defaults. Also clarifies misleading XML docs on the F1 2024 lap delta
fields and corrects the TimeTrialDataSet2024 assist flags, which the
spec defines as binary, to bool instead of the multi-valued
CarStatus/CarSetup enums.

+++ F1Server.Core/Packets/Data/F1-2026/LapPositions2026.cs
@@ -23,10 +23,10 @@ public class LapPositions2026 : ILapPositionsBase
     #region ILapPositionsBase
 
     /// <inheritdoc/>
-    public int NumberOfLaps { get; set; }
+    public ushort NumberOfLaps { get; set; }
 
     /// <inheritdoc/>
-    public int LapStartIndex { get; set; }
+    public ushort LapStartIndex { get; set; }
 
     /// <inheritdoc/>
     public int[,] CarPositionOnLaps { get; set; }
